move ResourceQueue
